modf.cpp

+++ test/doc/core/modf.cpp
@@ -4,7 +4,7 @@
 
 int main()
 {
-  eve::wide wf0{-0.0,  1.30, -1.3,  eve::inf(eve::as<double>()),
+  eve::wide wf0{-0.0f,  1.30f, -1.3f,  eve::inf(eve::as<float>()),
       0.0f, eve::nan(eve::as<float>()), 2.0f,  eve::prev(2.0f)};
 
    std::cout << "<- wf0                          = " << wf0 << "\n";
